+++ pc_IndustrialBuildingAI.cs
@@ -170,7 +170,7 @@ namespace RealCity
                         trade_tax = -trade_income * 0.04f;
                         Singleton<EconomyManager>.instance.AddPrivateIncome((int)trade_tax, ItemClass.Service.Industrial, data.Info.m_class.m_subService, data.Info.m_class.m_level, 111);
                     }
-                    comm_data.building_money[buildingID] = comm_data.building_money[buildingID] - (int)(trade_income + trade_tax);
+                    comm_data.building_money[buildingID] = (short)(comm_data.building_money[buildingID] - (int)(trade_income + trade_tax));
                     break;
                 case TransferManager.TransferReason.Food:
                     trade_tax = 0;
@@ -180,7 +180,7 @@ namespace RealCity
                         trade_tax = -trade_income * 0.04f;
                         Singleton<EconomyManager>.instance.AddPrivateIncome((int)trade_tax, ItemClass.Service.Industrial, data.Info.m_class.m_subService, data.Info.m_class.m_level, 111);
                     }
-                    comm_data.building_money[buildingID] = comm_data.building_money[buildingID] - (int)(trade_income + trade_tax);
+                    comm_data.building_money[buildingID] = (short)(comm_data.building_money[buildingID] - (int)(trade_income + trade_tax));
                     break;
                 case TransferManager.TransferReason.Petrol:
                     trade_tax = 0;
@@ -190,7 +190,7 @@ namespace RealCity
                         trade_tax = -trade_income * 0.1f;
                         Singleton<EconomyManager>.instance.AddPrivateIncome((int)trade_tax, ItemClass.Service.Industrial, data.Info.m_class.m_subService, data.Info.m_class.m_level, 111);
                     }
-                    comm_data.building_money[buildingID] = comm_data.building_money[buildingID] - (int)(trade_income + trade_tax);
+                    comm_data.building_money[buildingID] = (short)(comm_data.building_money[buildingID] - (int)(trade_income + trade_tax));
                     break;
                 case TransferManager.TransferReason.Coal:
                     trade_tax = 0;
@@ -200,7 +200,7 @@ namespace RealCity
                         trade_tax = -trade_income * 0.1f;
                         Singleton<EconomyManager>.instance.AddPrivateIncome((int)trade_tax, ItemClass.Service.Industrial, data.Info.m_class.m_subService, data.Info.m_class.m_level, 111);
                     }
-                    comm_data.building_money[buildingID] = comm_data.building_money[buildingID] - (int)(trade_income + trade_tax);
+                    comm_data.building_money[buildingID] = (short)(comm_data.building_money[buildingID] - (int)(trade_income + trade_tax));
                     break;
                 case TransferManager.TransferReason.Goods:
                     trade_tax = 0;
@@ -210,7 +210,7 @@ namespace RealCity
                         trade_tax = -trade_income * 0.1f;
                         Singleton<EconomyManager>.instance.AddPrivateIncome((int)trade_tax, ItemClass.Service.Industrial, data.Info.m_class.m_subService, data.Info.m_class.m_level, 111);
                     }
-                    comm_data.building_money[buildingID] = comm_data.building_money[buildingID] - (int)(trade_income + trade_tax);
+                    comm_data.building_money[buildingID] = (short)(comm_data.building_money[buildingID] - (int)(trade_income + trade_tax));
                     break;
             }
         }
